Update the main file
