--- rasterstats/src/rasterstats/__init__.py
@@ -1,15 +1,0 @@
-# isort: skip_file
-from rasterstats.main import gen_zonal_stats, raster_stats, zonal_stats
-from rasterstats.point import gen_point_query, point_query
-from rasterstats import cli
-from rasterstats._version import __version__
-
-__all__ = [
-    "__version__",
-    "gen_zonal_stats",
-    "gen_point_query",
-    "raster_stats",
-    "zonal_stats",
-    "point_query",
-    "cli",
-]
