--- rasterstats/src/rasterstats/_version.py
@@ -1,1 +1,0 @@
-__version__ = "0.19.0"
